+++ NEWS.md
@@ -1,3 +1,7 @@
+# crew.aws.batch 0.0.7.9000 (development)
+
+
+
 # crew.aws.batch 0.0.7
 
 * Send both cancellation and termination requests to end jobs.